==> 8.OnlinePlatformsData.ipynb <==
# Online Platforms Data - Microsoft Teams

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from matplotlib.dates import DateFormatter
from matplotlib.ticker import FuncFormatter

df = pd.read_csv('./OnlinePlatformsData/Data.csv')

df["Day"] = pd.to_datetime(df["Day"], dayfirst=True)

# Plot
plt.figure(figsize=(14, 6))

# Daily values
plt.plot(df["Day"], df["daily"], marker="o", label="Daily", color="tab:blue")

# Monthly values
plt.plot(df["Day"], df["monthly"], marker="s", label="Monthly", color="tab:orange")

# Format x-axis to show exact dates
plt.gca().xaxis.set_major_formatter(DateFormatter("%d %B %Y"))
#plt.gca().xaxis.set_major_formatter(DateFormatter("%d/%m/%Y"))
plt.xticks(df["Day"])  # Force all dates to be ticks
plt.xticks(rotation=45)
def millions(x, _):
    if x == 0:
        return "0"
    return f"{x / 1_000_000:.0f} million"
plt.gca().yaxis.set_major_formatter(FuncFormatter(millions))
# Labels and title
plt.title("Daily or Monthly Teams active users")
plt.xlabel("Report Date")
plt.ylabel("Teams active users")
plt.legend()
plt.grid(True)
plt.tight_layout()

plt.show()
